refactor: move function to separate list
