+++ clone-wars/src/js/data/EN_dataExpenseCategories.js
